+++ rust/foxglove/src/protobuf.rs
@@ -1,0 +1,277 @@
+use prost_types::field_descriptor_proto::Type as ProstFieldType;
+
+/// Serializes a Protocol Buffers FileDescriptorSet to a byte vector.
+///
+/// This function encodes the provided FileDescriptorSet message into its binary
+/// protobuf representation, which can be used for schema exchange and message
+/// type definitions in Foxglove.
+///
+/// # Arguments
+///
+/// * `file_descriptor_set` - A reference to the Protocol Buffers FileDescriptorSet to serialize
+///
+/// # Returns
+///
+/// A `Vec<u8>` containing the binary protobuf encoding of the FileDescriptorSet
+pub fn prost_file_descriptor_set_to_vec(
+    file_descriptor_set: &prost_types::FileDescriptorSet,
+) -> Vec<u8> {
+    use prost::Message;
+    file_descriptor_set.encode_to_vec()
+}
+
+/// The `ProtobufField` trait defines the interface for types that can be serialized to Protocol
+/// Buffer format.
+///
+/// This trait is automatically implemented for custom types when using the `#[derive(Loggable)]`
+/// attribute. It provides the necessary methods to serialize data according to Protocol Buffer
+/// encoding rules and generate appropriate Protocol Buffer schema information.
+///
+/// # Usage
+///
+/// This trait is typically implemented automatically by using the `#[derive(Loggable)]` attribute
+/// on your custom types:
+///
+/// ```rust
+/// #[derive(foxglove::Loggable)]
+/// struct MyMessage {
+///     number: u64,
+///     text: String,
+/// }
+/// ```
+pub trait ProtobufField {
+    /// Returns the Protocol Buffer field type that corresponds to this Rust type.
+    fn field_type() -> ProstFieldType;
+
+    /// Returns the Protocol Buffer wire type for this Rust type
+    fn wire_type() -> u32;
+
+    /// Writes a field with its tag (field number and wire type) to the buffer.
+    ///
+    /// The default implementation writes the tag followed by the field content.
+    fn write_tagged(&self, field_number: u32, buf: &mut impl bytes::BufMut) {
+        let tag = (field_number << 3) | Self::wire_type();
+        buf.put_u8(tag as u8);
+        self.write(buf);
+    }
+
+    /// Writes the field content to the output buffer according to Protocol Buffer encoding rules.
+    fn write(&self, buf: &mut impl bytes::BufMut);
+
+    /// Returns the type name for the type.
+    ///
+    /// For complex types (messages, enums) this should return the type name. For primitive types
+    /// this should return None (the default).
+    fn type_name() -> Option<String> {
+        None
+    }
+
+    /// If this trait is implemented on an Enum type, this returns the enum descriptor for the type.
+    fn enum_descriptor() -> Option<prost_types::EnumDescriptorProto> {
+        None
+    }
+
+    /// If this trait is implemented on a struct type, this returns the message descriptor for the type.
+    fn message_descriptor() -> Option<prost_types::DescriptorProto> {
+        None
+    }
+
+    /// Indicates the type represents a repeated field (like a Vec).
+    ///
+    /// By default, fields are not repeated.
+    fn repeating() -> bool {
+        false
+    }
+}
+
+// Implement ProtobufField for u64 that serializes the value as a varint in protobuf serialization
+impl ProtobufField for u64 {
+    fn field_type() -> ProstFieldType {
+        ProstFieldType::Uint64
+    }
+
+    fn wire_type() -> u32 {
+        0
+    }
+
+    fn write(&self, buf: &mut impl bytes::BufMut) {
+        let mut value = *self;
+        while value >= 0x80 {
+            buf.put_u8((value as u8) | 0x80);
+            value >>= 7;
+        }
+        buf.put_u8(value as u8);
+    }
+}
+
+impl ProtobufField for f32 {
+    fn field_type() -> ProstFieldType {
+        ProstFieldType::Float
+    }
+
+    fn wire_type() -> u32 {
+        5 // Float
+    }
+
+    fn write(&self, buf: &mut impl bytes::BufMut) {
+        buf.put_f32_le(*self);
+    }
+}
+
+impl ProtobufField for f64 {
+    fn field_type() -> ProstFieldType {
+        ProstFieldType::Double
+    }
+
+    fn wire_type() -> u32 {
+        1 // Double
+    }
+
+    fn write(&self, buf: &mut impl bytes::BufMut) {
+        buf.put_f64_le(*self);
+    }
+}
+
+// Implement ProtobufField for String that serializes the value in protobuf format
+impl ProtobufField for String {
+    fn field_type() -> ProstFieldType {
+        ProstFieldType::String
+    }
+
+    fn wire_type() -> u32 {
+        2 // Length-delimited
+    }
+
+    fn write(&self, buf: &mut impl bytes::BufMut) {
+        // Get the UTF-8 bytes of the string
+        let bytes = self.as_bytes();
+
+        // Write the length as a varint
+        let len = bytes.len();
+        let mut len_value = len as u64;
+        while len_value >= 0x80 {
+            buf.put_u8((len_value as u8) | 0x80);
+            len_value >>= 7;
+        }
+        buf.put_u8(len_value as u8);
+
+        // Write the string bytes
+        buf.put_slice(bytes);
+    }
+}
+
+// Implement ProtobufField for &str, which delegates to String's implementation
+impl ProtobufField for &str {
+    fn field_type() -> ProstFieldType {
+        <String as ProtobufField>::field_type()
+    }
+
+    fn wire_type() -> u32 {
+        <String as ProtobufField>::wire_type()
+    }
+
+    fn write(&self, buf: &mut impl bytes::BufMut) {
+        // Get the UTF-8 bytes of the string
+        let bytes = self.as_bytes();
+
+        // Write the length as a varint
+        let len = bytes.len();
+        let mut len_value = len as u64;
+        while len_value >= 0x80 {
+            buf.put_u8((len_value as u8) | 0x80);
+            len_value >>= 7;
+        }
+        buf.put_u8(len_value as u8);
+
+        // Write the string bytes
+        buf.put_slice(bytes);
+    }
+}
+
+// Implement ProtobufField for u32 that serializes the value as a varint in protobuf serialization
+impl ProtobufField for u32 {
+    fn field_type() -> ProstFieldType {
+        ProstFieldType::Uint32
+    }
+
+    fn wire_type() -> u32 {
+        0 // Varint
+    }
+
+    fn write(&self, buf: &mut impl bytes::BufMut) {
+        // For u32, we encode as a varint in the same way as u64
+        let mut value = *self as u64;
+        while value >= 0x80 {
+            buf.put_u8((value as u8) | 0x80);
+            value >>= 7;
+        }
+        buf.put_u8(value as u8);
+    }
+}
+
+// Implement ProtobufField for u16 that serializes the value as a varint in protobuf serialization
+impl ProtobufField for u16 {
+    fn field_type() -> ProstFieldType {
+        ProstFieldType::Uint32
+    }
+
+    fn wire_type() -> u32 {
+        0 // Varint
+    }
+
+    fn write(&self, buf: &mut impl bytes::BufMut) {
+        // For u16, we encode as a varint in the same way as u32
+        let mut value = *self as u64;
+        while value >= 0x80 {
+            buf.put_u8((value as u8) | 0x80);
+            value >>= 7;
+        }
+        buf.put_u8(value as u8);
+    }
+}
+
+// implement a protobuf field for any Vec<T> where T implements ProtobufField
+impl<T> ProtobufField for Vec<T>
+where
+    T: ProtobufField,
+{
+    fn field_type() -> ProstFieldType {
+        T::field_type()
+    }
+
+    fn wire_type() -> u32 {
+        2 // Length-delimited
+    }
+
+    fn write_tagged(&self, field_number: u32, buf: &mut impl bytes::BufMut) {
+        // non-packed repeated fields are encoded as a record for each element
+        // https://protobuf.dev/programming-guides/encoding/#optional
+        for value in self {
+            let wire_type = T::wire_type();
+
+            let tag = (field_number << 3) | wire_type;
+            buf.put_u8(tag as u8);
+
+            value.write(buf);
+        }
+    }
+
+    fn write(&self, _buf: &mut impl bytes::BufMut) {
+        panic!("Vec<T> should always be written using write_tagged");
+    }
+
+    fn repeating() -> bool {
+        true
+    }
+
+    fn message_descriptor() -> Option<prost_types::DescriptorProto> {
+        // The message descriptor of a vector is the message descriptor of the element type
+        // the "repeating" property is set on the field that is repeating rather than the message
+        // descriptor
+        T::message_descriptor()
+    }
+
+    fn type_name() -> Option<String> {
+        T::type_name()
+    }
+}
